examples/11-tetris/pyproject.toml
[project]
name = "vercel-tetris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "asyncpg>=0.31.0",
    "authlib>=1.6.6",
    "fastapi>=0.125.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "alembic>=1.17.2",
    "uvicorn>=0.38.0",
]
